+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 * Fix site delete error in k8s
 * Move docker builds from travis to github actions
+* Add email image header setting field to authentication tab (using filepond to upload image via the image server)
 
 ## v0.21.0
 * Update react admin